--- a/grpcutil/creds.go
+++ b/grpcutil/creds.go
@@ -8,8 +8,6 @@ import (
 )
 
 var tlsVersions = map[string]uint16{
-	"1.0": tls.VersionTLS10,
-	"1.1": tls.VersionTLS11,
 	"1.2": tls.VersionTLS12,
 	"1.3": tls.VersionTLS13,
 }