@fische 2021-11-09
The flag defaults now to TLS 1.3:
```
$ plz-out/bin/elan/elan --port 7777 -k ./key.pem -c ./cert.pem -s file://$(mktemp -d) --parallelism 20 --dir_cache_size 40960
$ nmap --script ssl-enum-ciphers -p 7777 localhost

Starting Nmap 7.60 ( https://nmap.org ) at 2021-11-09 13:52 GMT
Nmap scan report for localhost (127.0.0.1)
Host is up (0.000054s latency).
Other addresses for localhost (not scanned): ::1

PORT     STATE SERVICE
7777/tcp open  cbt

Nmap done: 1 IP address (1 host up) scanned in 0.13 seconds
```