@fische 2021-11-09
> Thanks!
> 
> YOu probably want to bump the VERSION file too to get a new release?

Oh right, yes, good idea